requirements.txt
# Core dependencies
anthropic>=0.20.0
playwright>=1.40.0
jinja2>=3.1.0
weasyprint>=60.0
sqlalchemy>=2.0.0
pyyaml>=6.0.1
click>=8.1.0
pydantic>=2.5.0
python-dateutil>=2.8.0

# Dashboard
streamlit>=1.28.0
plotly>=5.18.0
pandas>=2.1.0

# Utilities
tenacity>=8.2.0
rich>=13.7.0
python-dotenv>=1.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.1.0
mypy>=1.7.0
